refactor: rename account claims to accountClaims to prep for core expansion
We plan on merging the internal/user into internal/core, hence preparing distinction between account- and user claims.
